redundant files clean up
