--- los-urlopos/src/components/MockComponents/MockLogin.tsx
@@ -1,7 +1,0 @@
-import React, { FC } from "react";
-
-const MockLogin: FC = () => {
-  return <div>MockLogin - replace with actual Login Page</div>;
-};
-
-export default MockLogin;
