--- los-urlopos/src/components/MockComponents/MockUser.tsx
@@ -1,7 +1,0 @@
-import React, { FC } from "react";
-
-const MockUser: FC = () => {
-  return <div>MockUser - replace with actual user page</div>;
-};
-
-export default MockUser;
