--- los-urlopos/src/components/MockComponents/MockUserPanel.tsx
@@ -1,7 +1,0 @@
-import React, { FC } from "react";
-
-const MockUserPanel: FC = () => {
-  return <div>MockUserPanel - replace with actual userPanel</div>;
-};
-
-export default MockUserPanel;
